Refactor duplicated code into batch and godot helper files. Cleanup training scripts
