+++ scripts/training/ppo_individual.py
@@ -18,27 +18,28 @@ experiences only.
 Branched from CleanRL ppo_continuous_action.py at https://github.com/vwxyzjn/cleanrl/blob/master/cleanrl/ppo_continuous_action.py
 '''
 
+import os
 import random
 import traceback
 import time
-from dataclasses import dataclass
-from typing import Optional
 import timeit
-
-import numpy as np
 import torch
+import tyro
+import numpy as np
 import torch.nn as nn
 import torch.optim as optim
-import tyro
+from dataclasses import dataclass
+from typing import Optional
 from torch.utils.tensorboard import SummaryWriter
-from godot_rl.wrappers.petting_zoo_wrapper import GDRLPettingZooEnv
 
 import sys, os
 parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(parent_dir)
 
-from normalize import normalize_obs
+from batch_helpers import batchify_obs, batchify, unbatchify, load_state_dicts
 from agents.jestel_agent import Agent
+from godot import make_env
+
 
 @dataclass
 class Args:
@@ -99,55 +100,6 @@ class Args:
     num_updates: int = 0
     """the number of times the policy will update (computed in runtime)"""
 
-def make_env(time_scale, file_path, no_graphics, seed):
-    config = {
-        "env_path": file_path,
-        "show_window": no_graphics,
-        "speedup": time_scale
-    }
-    return GDRLPettingZooEnv(config=config, seed=seed)
-
-def batchify_obs(obs, device):
-    """Converts dict style observations to a multi-dimensional torch array."""
-    batched_obs = np.array([]).reshape(0, Agent.stacked_observation_size())
-    for agent_obs_dict_values in obs.values():
-        agent_obs = (np.array(list(agent_obs_dict_values.values())))
-
-        assert (agent_obs.shape[1] == Agent.stacked_observation_size()), "Received incorrect Agent stacked observation size"
-        batched_obs = np.concatenate((batched_obs, agent_obs), axis=0)
-    
-    batched_obs = normalize_obs(batched_obs, Agent.get_observation_range())
-    batched_obs = torch.tensor(batched_obs, dtype=torch.float32).to(device)
-    return batched_obs
-
-def batchify(x, device):
-    """Converts dict style returns and dones to batch of torch arrays."""
-    x = np.stack([x[a] for a in x], axis=0) # convert to list of np arrays
-    x = torch.tensor(x).to(device)
-    return x
-
-def unbatchify(x, env):
-    """Converts torch array of returns and dones to dict style."""
-    x = x.cpu().numpy()
-    x = {a: np.expand_dims(x[i], 0) for i, a in enumerate(env.possible_agents)}
-    return x
-
-def load_state_dicts(model_path, agents):
-    """Loads the state dict of each agent with a model filepath or filepath to 
-    a folder containing the models"""
-    if (args.model_path):
-        if (args.model_path.endswith('/')):
-            print("Loading pre-existing models inside directory [" + args.model_path + "].")
-
-            model_paths = [(args.model_path + file) for file in os.listdir(args.model_path) if file.endswith('.cleanrl_model')]
-            assert (len(model_paths) == len(agents)), "Found " + str(len(model_paths)) + " agent models but require " + str(len(agents)) + " for this environment."
-
-            for model_path, agent in zip(model_paths, agents):
-                agent.load_state_dict(torch.load(model_path, map_location=device, weights_only=False))
-        else:
-            print("Loading pre-existing model [" + args.model_path + "]")
-            for agent in agents:
-                agent.load_state_dict(torch.load(args.model_path, map_location=device, weights_only=False))
 
 if __name__ == "__main__":
     args = tyro.cli(Args)
@@ -172,7 +124,7 @@ if __name__ == "__main__":
 
     num_agents = len(env.possible_agents)
     agents = [Agent(observation_space, action_space).to(device) for i in range(num_agents)] 
-    load_state_dicts(args.model_path, agents)
+    load_state_dicts(args.model_path, agents, device)
 
     optimizers = [optim.Adam(agent.parameters(), lr=args.learning_rate, eps=1e-5) for agent in agents]
 
@@ -194,7 +146,7 @@ if __name__ == "__main__":
     episode_start_step = 0
     total_episodic_reward = 0
     start_time = time.time()
-    next_obs = batchify_obs(env.reset()[0], device)
+    next_obs = batchify_obs(env.reset()[0], device, Agent)
     next_done = torch.zeros(num_agents).to(device)
 
     try:
@@ -228,7 +180,7 @@ if __name__ == "__main__":
                 # Step the environment with the batched policy actions
                 next_obs_unbatched, reward, next_done, _, infos = env.step(unbatchify(new_actions, env))
 
-                next_obs = batchify_obs(next_obs_unbatched, device)
+                next_obs = batchify_obs(next_obs_unbatched, device, Agent)
                 next_done = batchify(next_done, device).long()
 
                 rewards[step] = batchify(reward, device).view(-1)
@@ -244,7 +196,7 @@ if __name__ == "__main__":
                     writer.add_scalar("charts/episodic_length", (global_step - episode_start_step), global_step)
                     print(episodic_msg)
 
-                    next_obs = batchify_obs(env.reset()[0], device)
+                    next_obs = batchify_obs(env.reset()[0], device, Agent)
                     total_episodic_reward = 0
                     episode_start_step = global_step
